Fix the navbar background issue, and move to unminified JS
This fixes an issue on mobile, where the navbar would remain transparent at the top of the page, when the menu is expanded. This caused text to overlap on each other, which was not pretty. This is because the navbar is transparent until you scroll further down the page. For now, my fix is to identify that we are still near the top of the page by adding an extra class to the navbar element, which is then checked in the nav open/close functions. This then allows us to decide whether to make the navbar solid, or transparent, without breaking it after you've already scrolled.

This also moves all pages to the unminified JS, as it's much easier to work with until I setup automated minify/build processes.