Corrected hyperlink to the second Downloads button
Hi Team,

Originally opened here: https://github.com/jellyfin/jellyfin.github.io/pull/42
The second button to Downloads was hyperlinked to "/docs/general/#getting-started", which was throwing a 404. I saw that the first button on that page was hyperlinked to "/downloads". I am opening a PR to change the 2nd hyperlink to "/downloads" instead.